Staging to Master for Goat Tree release

+++ tola_management/countryadmin.py
@@ -1,13 +1,15 @@
-import json
 from collections import OrderedDict
+import json
 from django.db import transaction
 from django.db.models import Q
 from rest_framework import viewsets
 from rest_framework.response import Response
+from rest_framework import status
 
 from rest_framework import serializers
 from rest_framework import permissions
 from rest_framework import pagination
+from rest_framework.decorators import action
 
 
 from workflow.models import (
@@ -21,6 +23,8 @@ from indicators.models import (
     DisaggregationType,
     DisaggregationLabel,
 )
+
+from tola_management.models import CountryAdminAuditLog
 
 from tola_management.permissions import (
     HasCountryAdminAccess,
@@ -116,6 +120,14 @@ class CountryAdminViewSet(viewsets.ModelViewSet):
 
         return queryset.distinct()
 
+    @action(detail=True, methods=['GET'])
+    def history(self, request, pk=None):
+        country = Country.objects.get(pk=pk)
+        queryset = CountryAdminAuditLog.objects.filter(
+            country=country).select_related("admin_user", "disaggregation_type").order_by('-date')
+        serializer = CountryAdminAuditLogSerializer(queryset, many=True)
+        return Response(serializer.data)
+
 
 class CountryObjectiveSerializer(serializers.ModelSerializer):
     #id = serializers.IntegerField(allow_null=True, required=False)
@@ -156,16 +168,18 @@ class CountryObjectiveViewset(viewsets.ModelViewSet):
 class NestedDisaggregationLabelSerializer(serializers.ModelSerializer):
     id = serializers.IntegerField(required=False, allow_null=True)
     label = serializers.CharField(required=True)
+    customsort = serializers.IntegerField(required=True)
     class Meta:
         model = DisaggregationLabel
         fields = (
             'id',
             'label',
+            'customsort'
         )
 
     def to_representation(self, disaggregation_label):
         ret = super(NestedDisaggregationLabelSerializer, self).to_representation(disaggregation_label)
-        ret['in_use'] = disaggregation_label.disaggregationvalue_set.exists()
+        ret['in_use'] = disaggregation_label.disaggregatedvalue_set.filter(value__gt=0).exists()
         return ret
 
     def to_internal_value(self, data):
@@ -173,9 +187,10 @@ class NestedDisaggregationLabelSerializer(serializers.ModelSerializer):
             data.pop('id')
         validated_data = super(NestedDisaggregationLabelSerializer, self).to_internal_value(data)
         instance = None
-        if validated_data.get(id):
-            instance = DisaggregationLabel.objects.get(pk=validated_data.id)
-            instance.update(**validated_data)
+        if validated_data.get('id'):
+            instance = DisaggregationLabel.objects.get(pk=validated_data.pop('id'))
+            for field, value in validated_data.items():
+                setattr(instance, field, value)
         else:
             instance = DisaggregationLabel(**validated_data)
         return instance
@@ -188,6 +203,8 @@ class CountryDisaggregationSerializer(serializers.ModelSerializer):
         required=False,
         many=True,
     )
+    has_indicators = serializers.BooleanField(read_only=True)
+    is_archived = serializers.BooleanField(required=False)
 
     class Meta:
         model = DisaggregationType
@@ -196,29 +213,61 @@ class CountryDisaggregationSerializer(serializers.ModelSerializer):
             'country',
             'disaggregation_type',
             'labels',
+            'has_indicators',
+            'is_archived',
+            'selected_by_default'
         )
 
     @transaction.atomic
     def update(self, instance, validated_data):
-        updated_label_data = validated_data.pop('disaggregationlabel_set')
-        current_labels = [label for label in instance.disaggregationlabel_set.all()]
-        removed_labels = [label for label in current_labels if label not in updated_label_data]
-        new_labels = [label for label in updated_label_data if label not in current_labels]
-        for label in new_labels:
-            label.disaggregation_type = instance
-            label.save()
-        for label in removed_labels:
-            label.delete()
+        previous_entry = instance.logged_fields
+        change_type = "country_disaggregation_updated"
+        if previous_entry['is_archived'] and not validated_data['is_archived']:
+            change_type = "country_disaggregation_unarchived"
+        updated_label_data = validated_data.pop('disaggregationlabel_set', None)
+        if not self.partial or updated_label_data is not None:
+            current_labels = [label for label in instance.disaggregationlabel_set.all()]
+            removed_labels = [label for label in current_labels if label not in updated_label_data]
+            new_labels = [label for label in updated_label_data if label not in current_labels]
+            for label in new_labels:
+                label.disaggregation_type = instance
+                label.save()
+            for label in updated_label_data:
+                label.save()
+            for label in removed_labels:
+                label.delete()
         updated_instance = super(CountryDisaggregationSerializer, self).update(instance, validated_data)
+
+        if updated_instance.logged_fields != previous_entry:
+            CountryAdminAuditLog.objects.create(
+                admin_user=self.context['tola_user'],
+                country = instance.country,
+                disaggregation_type=instance,
+                change_type=change_type,
+                previous_entry=json.dumps(previous_entry),
+                new_entry=json.dumps(updated_instance.logged_fields),
+            )
+
         return updated_instance
 
     @transaction.atomic
     def create(self, validated_data):
+
         labels = validated_data.pop('disaggregationlabel_set')
         instance = super(CountryDisaggregationSerializer, self).create(validated_data)
         for label in labels:
             label.disaggregation_type = instance
             label.save()
+
+        CountryAdminAuditLog.objects.create(
+            admin_user=self.context['tola_user'],
+            country = instance.country,
+            disaggregation_type=instance,
+            change_type="country_disaggregation_created",
+            previous_entry={},
+            new_entry=json.dumps(instance.logged_fields),
+        )
+
         return instance
 
 class CountryDisaggregationViewSet(viewsets.ModelViewSet):
@@ -234,3 +283,69 @@ class CountryDisaggregationViewSet(viewsets.ModelViewSet):
             queryset = queryset.filter(country__pk=countryFilter)
 
         return queryset
+
+    def destroy(self, request, pk=None):
+        disaggregation = DisaggregationType.objects.get(pk=pk)
+        previous_entry = disaggregation.logged_fields
+        previous_country = disaggregation.country
+        if disaggregation.has_indicators:
+            disaggregation.is_archived = True
+            disaggregation.save()
+            # TODO: try catch here in case of bad model / unarchivable?  Test whether archive is the right action?
+            CountryAdminAuditLog.objects.create(
+                admin_user=self.request.user.tola_user,
+                country=disaggregation.country,
+                disaggregation_type=disaggregation,
+                change_type="country_disaggregation_archived",
+                previous_entry=json.dumps(previous_entry),
+                new_entry=json.dumps(disaggregation.logged_fields),
+            )
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            destroyed = super().destroy(request, pk)
+            CountryAdminAuditLog.objects.create(
+                admin_user=self.request.user.tola_user,
+                country=previous_country,
+                disaggregation_type=None,
+                change_type="country_disaggregation_deleted",
+                previous_entry=json.dumps(previous_entry),
+                new_entry={},
+            )
+
+            return destroyed
+
+    def get_serializer_context(self):
+        context = super(CountryDisaggregationViewSet, self).get_serializer_context()
+        context['tola_user'] = self.request.user.tola_user
+        return context
+
+
+class CountryAdminAuditLogSerializer(serializers.ModelSerializer):
+    date = serializers.DateTimeField(format="%Y-%m-%d %H:%M:%S")
+    admin_user = serializers.CharField(source="admin_user.name")
+    disaggregation_type = serializers.SerializerMethodField()
+    disaggregation_type_name = serializers.SerializerMethodField()
+
+    class Meta:
+        model = CountryAdminAuditLog
+        fields = [
+            "id",
+            "date",
+            "admin_user",
+            "disaggregation_type",
+            "disaggregation_type_name",
+            "pretty_change_type",
+            "diff_list"
+        ]
+
+    def get_disaggregation_type(self, obj):
+        if obj.disaggregation_type:
+            return obj.disaggregation_type.pk
+        else:
+            return ""
+
+    def get_disaggregation_type_name(self, obj):
+        if obj.disaggregation_type:
+            return obj.disaggregation_type.disaggregation_type
+        else:
+            return ""
